[Tetris #1] BREAKING: Moving all files into subdirectories include/ and src/
This is an early attempt at organizing the project structure into 3 main
directories (See #1, #2)

However, header file includes are all messed up so many CMake configs
and includes would need to be rewritten
