+++ src/app/main.cpp
@@ -1,0 +1,10 @@
+#include <SDL2/SDL.h>
+
+#include "Game/Tetris.hpp"
+
+int main (int argc, char **argv)
+{
+	Tetris tetris;
+
+	return 0;
+}
