+++ src/core/include/Coordinate.hpp
