+++ src/core/include/Variables.hpp
