+++ src/core/src/Coordinate.cpp
