+++ src/core/src/Drop.cpp
