+++ src/core/src/Piece.cpp
