+++ src/core/src/Rotate.cpp
