+++ src/render/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(Render 
+	Renderer.cpp
+)
