+++ src/render/include/Core.hpp
