+++ src/render/src/Core.cpp
